+++ app/src/main/java/com/legue/axel/bankingapp/database/model/Ingredient.java
@@ -16,14 +16,14 @@ public class Ingredient {
 
     @PrimaryKey(autoGenerate = true)
     private int id;
-    private int quantity;
+    private double quantity;
     private String measure;
     @SerializedName("ingredient")
     private String name;
     private int recipeId;
 
 
-    public Ingredient(int id, int quantity, String measure, String name, int recipeId) {
+    public Ingredient(int id, double quantity, String measure, String name, int recipeId) {
         this.id = id;
         this.quantity = quantity;
         this.measure = measure;
@@ -39,11 +39,11 @@ public class Ingredient {
         this.id = id;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
